fix: Task 1レビュー指摘（バリデータテストの逐語複製解消・XMLコメント補完）
BeltConveyorRoleTestが非ベルト型/コスト免除のバリデータテスト2件を
BeltConveyorFamilyTestのヘルパー(LoadBlocksJson/FindBlock)と逐語複製していたため、
バリデータ系テストの本籍であるBeltConveyorFamilyTest側へ移設しRoleTest側の
System.IO/Newtonsoft.Json.Linq/Core.Master.Validator依存を落とした(Important #1)。

あわせてTryGetRole失敗時のout値が有効なStraightを返す点、
ResolveSlopeRepresentativeGuidの2回master解決という非自明な経路に
JP/EN 2行コメントを補完した(Minor #1, #3)。
